finished Yao protocol
git-svn-id: svn://localhost/repos@122 16ffc253-0420-0410-a4db-cfc06cdcecfe

+++ sfe/C-sfe/src/sillylib/silly.cpp
@@ -171,7 +171,7 @@ string silly::misc::string_printf(const char *fmt, ...) {
 		n = vsnprintf((char*)&buf[0], buf.size(), fmt, ap);
 		va_end(ap);
 		/* If that worked, return the string. */
-		if (n > -1 && n < buf.size())
+		if (n > -1 && uint(n) < buf.size())
 			return string((char*)&buf[0]);
 		/* Else try again with more space. */
 		if (n > -1)    /* glibc 2.1 */
